add password

+++ myApp.py
@@ -1,4 +1,6 @@
 import streamlit as st
+st.set_page_config(page_title = "TNBS Spooky Spikes Online", page_icon = "tnbs_logo.png")
+
 import neo.io
 import numpy as np
 import plotly.graph_objects as go
@@ -633,15 +635,45 @@ def main():
     st.sidebar.write(name11, val11)
 
 st.title('Spooky Spikes Online')
-    
-f = st.file_uploader('Select smr file to upload','smr',False)
-#st.write(os.getcwd())
 
-if f is not None:
-    filename = upload_smr_file(f)
-    main()
+def check_password():
+    """Returns `True` if the user had the correct password."""
 
-    #st.write(os.listdir())
+    def password_entered():
+        """Checks whether a password entered by the user is correct."""
+        if st.session_state["password"] == st.secrets["password"]:
+            st.session_state["password_correct"] = True
+            del st.session_state["password"]  # don't store password
+        else:
+            st.session_state["password_correct"] = False
+
+    if "password_correct" not in st.session_state:
+        # First run, show input for password.
+        st.text_input(
+            "Password", type="password", on_change=password_entered, key="password"
+        )
+        return False
+    elif not st.session_state["password_correct"]:
+        # Password not correct, show input + error.
+        st.text_input(
+            "Password", type="password", on_change=password_entered, key="password"
+        )
+        st.error("Password incorrect")
+        return False
+    else:
+        # Password correct.
+        return True
+
+if check_password():
     
-    if os.path.isfile(filename):
-        os.remove(filename)   
+    f = st.file_uploader('Select smr file to upload','smr',False)
+    #st.write(os.getcwd())
+
+    if f is not None:
+        filename = upload_smr_file(f)
+        main()
+
+        #st.write(os.listdir())
+        
+        if os.path.isfile(filename):
+            os.remove(filename)
